Refactor: Minor fixes from IDE type checking
